% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handlers.R
\name{choose_handler_for_file}
\alias{choose_handler_for_file}
\title{Choose an appropriate handler for a file path}
\usage{
choose_handler_for_file(path)
}
\arguments{
\item{path}{File path}
}
\value{
A function(path, max_chars, ...) -> list(...) (the handler)
}
\description{
Choose an appropriate handler for a file path
}
