% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentoR.R
\name{create_readme_text_fixed}
\alias{create_readme_text_fixed}
\title{Create README.md content (use SUMMARY section)}
\usage{
create_readme_text_fixed(
  project_path,
  project_summary = NULL,
  file_summaries = list(),
  readme_title = NULL
)
}
\description{
Create README.md content (use SUMMARY section)
}
