% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentoR.R
\name{create_single_documentation}
\alias{create_single_documentation}
\title{Create a single combined documentation markdown file}
\usage{
create_single_documentation(
  summaries,
  output_file = "documentation/documentation.md",
  include_frontmatter = TRUE
)
}
\description{
Create a single combined documentation markdown file
}
