% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentoR.R
\name{document_project}
\alias{document_project}
\title{Document an R project (single_output supported)}
\usage{
document_project(
  path = ".",
  chat_fn = NULL,
  output_dir = NULL,
  output_base = getwd(),
  max_chars = 16000,
  verbose = TRUE,
  generate_docs = TRUE,
  generate_readme = TRUE,
  overwrite_readme = FALSE,
  single_output = FALSE,
  single_output_file = "documentation/documentation.md",
  exclude_file = NULL,
  exclude_dirs = NULL,
  exclude_patterns = NULL,
  change_only = FALSE
)
}
\arguments{
\item{output_base}{"path" | "project" | path string. If omitted, defaults to getwd().}
}
\description{
Document an R project (single_output supported)
}
