% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handlers.R
\docType{data}
\name{FILE_HANDLERS}
\alias{FILE_HANDLERS}
\title{Registry of file handlers (extension -> function)
Add or override entries as needed.}
\format{
An object of class \code{list} of length 22.
}
\usage{
FILE_HANDLERS
}
\description{
Registry of file handlers (extension -> function)
Add or override entries as needed.
}
\keyword{datasets}
