% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentoR.R
\name{file_summary_prompt}
\alias{file_summary_prompt}
\title{Default prompt template for per-file summarization}
\usage{
file_summary_prompt(file_path, file_content, meta = list())
}
\description{
Default prompt template for per-file summarization
}
