% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentoR.R
\name{read_file_snippet}
\alias{read_file_snippet}
\title{Read a file safely into a truncated string}
\usage{
read_file_snippet(file, max_chars = 16000)
}
\description{
Read a file safely into a truncated string
}
