% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentoR.R
\name{summarize_file}
\alias{summarize_file}
\title{Summarize a single file using a user-supplied chat function and file handler}
\usage{
summarize_file(file, chat_fn, max_chars = 16000, verbose = TRUE)
}
\description{
Summarize a single file using a user-supplied chat function and file handler
}
